Feat: Category 패키지 충돌 수정

+++ src/main/java/com/shoux_kream/item/dto/response/ItemResponse.java
@@ -14,7 +14,7 @@ public record ItemResponse(
         String detailDescription,
         String imageKey,
         int inventory,
-        double price,
+        int price,
         String searchKeywords
 ) {
     public static ItemResponse fromEntity(Item item) {
